refactor: 자료형 통일

+++ src/main/java/cart/dto/request/ProductRequest.java
@@ -15,9 +15,9 @@ public class ProductRequest {
     @NotNull(message = "가격은 비어있을 수 없습니다.")
     @Min(value = 0, message = "가격은 {min} 이상이여야 합니다.")
     @Max(value = 1_000_000_000, message = "가격은 {max} 이하여야 합니다.")
-    private final long price;
+    private final Integer price;
 
-    public ProductRequest(String image, String name, long price) {
+    public ProductRequest(String image, String name, Integer price) {
         this.image = image;
         this.name = name;
         this.price = price;
@@ -31,7 +31,7 @@ public class ProductRequest {
         return name;
     }
 
-    public long getPrice() {
+    public Integer getPrice() {
         return price;
     }
 }
