+++ src/main/java/cart/dto/response/ProductResponse.java
@@ -4,9 +4,9 @@ public class ProductResponse {
     private final Long id;
     private final String name;
     private final String image;
-    private final long price;
+    private final Integer price;
 
-    public ProductResponse(Long id, String name, String image, long price) {
+    public ProductResponse(Long id, String name, String image, Integer price) {
         this.id = id;
         this.name = name;
         this.image = image;
@@ -25,7 +25,7 @@ public class ProductResponse {
         return image;
     }
 
-    public long getPrice() {
+    public Integer getPrice() {
         return price;
     }
 }
